+++ test/common/http/utility_test.cc
@@ -1,7 +1,7 @@
 #include <cstdint>
 #include <string>
 
-#include "envoy/data/core/v2alpha/local_reply_body.pb.h"
+#include "envoy/data/core/v2alpha/local_reply.pb.h"
 #include "envoy/api/v2/core/protocol.pb.h"
 #include "envoy/api/v2/core/protocol.pb.validate.h"
 
@@ -548,13 +548,12 @@ TEST(HttpUtility, SendLocalReplyJsonConntentTypeRequest) {
       }));
   EXPECT_CALL(callbacks, encodeData(_, true))
       .WillOnce(Invoke([&](Buffer::Instance& data, bool) -> void {
-        envoy::data::core::v2alpha::LocalReplyBody local_reply_body;
-        local_reply_body.set_body("large");
-        EXPECT_EQ(MessageUtil::getJsonStringFromMessage(local_reply_body, true, true),
-                  data.toString());
+        envoy::data::core::v2alpha::LocalReply local_reply;
+        local_reply.set_body("large");
+        EXPECT_EQ(MessageUtil::getJsonStringFromMessage(local_reply, true, true), data.toString());
       }));
   Utility::LocalReplyInfo info;
-  info.has_json_content_type = true;
+  info.accept_json_type = true;
   Utility::sendLocalReply(info, callbacks, is_reset, Http::Code::PayloadTooLarge, "large",
                           absl::nullopt);
 }